refactor: change where some JK methods live from MeasureIABase to MeasureJackknife
