+++ src/measureia/measure_IA_base.py
@@ -1,7 +1,6 @@
 import math
 import numpy as np
 import h5py
-from kmeans_radec import kmeans_sample
 from scipy.special import lpmn
 from .write_data import write_dataset_hdf5, create_group_hdf5
 from .Sim_info import SimInfo
@@ -377,117 +376,6 @@ class MeasureIABase(SimInfo):
 				del setdiff
 		return diff
 
-	def _get_jackknife_region_indices(self, masks, L_subboxes):
-		"""
-		Split the box in L_subboxes^3 subboxes and return indices of which subbox objects are in for position and
-		shape sample.
-
-		Parameters
-		----------
-		masks: dict or NoneType
-			Input in methods in MeasureIABox that masks the input data dictionary.
-		L_subboxes: int
-			Number of subboxes on one side of the box. L_subboxes^3 is the total number of jackknife realisations.
-
-		Returns
-		-------
-		ndarrays
-			indices of jackknife region of position sample and indices of jackknife region of shape sample
-
-		"""
-		if masks == None:
-			positions = self.data["Position"]
-			positions_shape_sample = self.data["Position_shape_sample"]
-		else:
-			positions = self.data["Position"][masks["Position"]]
-			positions_shape_sample = self.data["Position_shape_sample"][masks["Position_shape_sample"]]
-		L_sub = self.L_0p5 * 2.0 / L_subboxes
-		jackknife_region_indices_pos = np.zeros(len(positions))
-		jackknife_region_indices_shape = np.zeros(len(positions_shape_sample))
-		num_box = 0
-		for i in np.arange(0, L_subboxes):
-			for j in np.arange(0, L_subboxes):
-				for k in np.arange(0, L_subboxes):
-					x_bounds = [i * L_sub, (i + 1) * L_sub]
-					y_bounds = [j * L_sub, (j + 1) * L_sub]
-					z_bounds = [k * L_sub, (k + 1) * L_sub]
-					x_mask = (positions[:, 0] > x_bounds[0]) * (positions[:, 0] < x_bounds[1])
-					y_mask = (positions[:, 1] > y_bounds[0]) * (positions[:, 1] < y_bounds[1])
-					z_mask = (positions[:, 2] > z_bounds[0]) * (positions[:, 2] < z_bounds[1])
-					x_mask_shape = (positions_shape_sample[:, 0] > x_bounds[0]) * (
-							positions_shape_sample[:, 0] < x_bounds[1])
-					y_mask_shape = (positions_shape_sample[:, 1] > y_bounds[0]) * (
-							positions_shape_sample[:, 1] < y_bounds[1])
-					z_mask_shape = (positions_shape_sample[:, 2] > z_bounds[0]) * (
-							positions_shape_sample[:, 2] < z_bounds[1])
-					mask_position = x_mask * y_mask * z_mask  # mask that is True for all positions in the subbox
-					mask_shape = x_mask_shape * y_mask_shape * z_mask_shape  # mask that is True for all positions not in the subbox
-					jackknife_region_indices_pos[mask_position] = num_box
-					jackknife_region_indices_shape[mask_shape] = num_box
-					num_box += 1
-		return np.array(jackknife_region_indices_pos, dtype=int), np.array(jackknife_region_indices_shape, dtype=int)
-
-	def _combine_jackknife_information(self, dataset_name, jk_group_name, corr_group, num_box, return_output=False):
-		"""
-		Combine jackknife realisations into a covariance matrix.
-
-		Parameters
-		----------
-		dataset_name: str
-			Name of the dataset in the output file.
-		jk_group_name: str
-			Name of the subgroup in the output file where the jackknife realisations are saved.
-		corr_group: list of str
-			Name of the subgroups in the output file denoting the correlation (e.g. w_g_plus, multipoles_gg etc).
-		num_box: int
-			Number of jackknife realisations.
-		return_output: bool, optional
-			When True, returns output, otherwise saves to output file.
-
-		Returns
-		-------
-		list of ndarrays
-			list of covariances for each entry in corr_group and list of standard deviations for each entry in corr_group
-
-		"""
-		covs, stds = [], []
-		for d in np.arange(0, len(corr_group)):
-			data_file = h5py.File(self.output_file_name, "a")
-			group_multipoles = data_file[f"{self.snap_group}/{corr_group[d]}/{jk_group_name}/"]
-			# calculating mean of the datavectors
-			mean_multipoles = np.zeros(self.num_bins_r)
-			for b in np.arange(0, num_box):
-				mean_multipoles += group_multipoles[dataset_name + "_" + str(b)][:]
-			mean_multipoles /= num_box
-
-			# calculation the covariance matrix (multipoles) and the standard deviation (sqrt of diag of cov)
-			cov = np.zeros((self.num_bins_r, self.num_bins_r))
-			std = np.zeros(self.num_bins_r)
-			for b in np.arange(0, num_box):
-				std += (group_multipoles[dataset_name + "_" + str(b)][:] - mean_multipoles) ** 2
-				for i in np.arange(self.num_bins_r):
-					cov[:, i] += (group_multipoles[dataset_name + "_" + str(b)][:] - mean_multipoles) * (
-							group_multipoles[dataset_name + "_" + str(b)][i] - mean_multipoles[i]
-					)
-			std *= (num_box - 1) / num_box  # see Singh 2023
-			std = np.sqrt(std)  # size of errorbars
-			cov *= (num_box - 1) / num_box  # cov not sqrt so to get std, sqrt of diag would need to be taken
-			data_file.close()
-			if return_output:
-				covs.append(cov)
-				stds.append(std)
-			else:
-				output_file = h5py.File(self.output_file_name, "a")
-				group_multipoles = create_group_hdf5(output_file, f"{self.snap_group}/" + corr_group[d])
-				write_dataset_hdf5(group_multipoles, dataset_name + "_mean_" + str(num_box), data=mean_multipoles)
-				write_dataset_hdf5(group_multipoles, dataset_name + "_jackknife_" + str(num_box), data=std)
-				write_dataset_hdf5(group_multipoles, dataset_name + "_jackknife_cov_" + str(num_box), data=cov)
-				output_file.close()
-		if return_output:
-			return covs, stds
-		else:
-			return
-
 	def _measure_w_g_i(self, dataset_name, corr_type="both", return_output=False, jk_group_name=""):
 		"""Measures w_gg or w_g+ for a given xi_gi dataset that has been calculated with the _measure_xi_rp_pi_sims
 		methods. Integrates over pi bins via sum * dpi. Stores rp, and w_gg or w_g+.
@@ -706,64 +594,7 @@ class MeasureIABase(SimInfo):
 		output_file.close()
 		return
 
-	def assign_jackknife_patches(self, data, randoms_data, num_jk):
-		"""Assigns jackknife patches to data and randoms given a number of patches.
-		Based on https://github.com/esheldon/kmeans_radec
 
-		Parameters
-		----------
-		data : dict
-			Dictionary containing position and shape sample data. Keywords: "RA", "DEC", "RA_shape_sample",
-			"DEC_shape_sample"
-		randoms_data : dict
-			Dictionary containing position and shape sample data of randoms. Keywords: "RA", "DEC", "RA_shape_sample",
-			"DEC_shape_sample"
-		num_jk : int
-			Number of jackknife patches
-
-		Returns
-		-------
-		dict
-			Dictionary with patch numbers for each sample. Keywords: 'position', 'shape', 'randoms_position',
-			'randoms_shape'
-
-		"""
-
-		jk_patches = {}
-
-		# Read the randoms file from which the jackknife regions will be created
-		RA = randoms_data['RA']
-		DEC = randoms_data['DEC']
-
-		# Define a number of jaccknife regions and find their centres using kmans
-		X = np.column_stack((RA, DEC))
-		km = kmeans_sample(X, num_jk, maxiter=100, tol=1.0e-5)
-		jk_labels = km.labels
-
-		jk_patches['randoms_position'] = jk_labels
-
-		RA = randoms_data['RA_shape_sample']
-		DEC = randoms_data['DEC_shape_sample']
-		X2 = np.column_stack((RA, DEC))
-		jk_labels = km.find_nearest(X2)
-
-		jk_patches['randoms_shape'] = jk_labels
-
-		RA = data['RA']
-		DEC = data['DEC']
-		X2 = np.column_stack((RA, DEC))
-		jk_labels = km.find_nearest(X2)
-
-		jk_patches['position'] = jk_labels
-
-		RA = data['RA_shape_sample']
-		DEC = data['DEC_shape_sample']
-		X2 = np.column_stack((RA, DEC))
-		jk_labels = km.find_nearest(X2)
-
-		jk_patches['shape'] = jk_labels
-
-		return jk_patches
 
 
 if __name__ == "__main__":
